requirements.txt
pandas
numpy
scikit-learn
Flask 
openai
python-dotenv
google-generativeai